Integrate Launchable into make btest
