+++ .github/workflows/rjit.yml
@@ -67,6 +67,8 @@ jobs:
           srcdir: src
           builddir: build
           makeup: true
+          # Set fetch-depth: 10 so that Launchable can receive commits information.
+          fetch-depth: 10
 
       - name: Run configure
         env:
@@ -77,13 +79,27 @@ jobs:
 
       - run: $SETARCH make
 
+      - name: Set up Launchable
+        uses: ./.github/actions/launchable/setup
+        with:
+          os: ubuntu-22.04
+          test-task: test
+          launchable-token: ${{ secrets.LAUNCHABLE_TOKEN }}
+          builddir: build
+          srcdir: src
+          launchable-workspace: ruby-make-btest
+          test-opts: ${{ matrix.run_opts }}
+        continue-on-error: true
+
       - name: make test
         run: |
           $SETARCH make -s test RUN_OPTS="$RUN_OPTS"
         timeout-minutes: 30
         env:
           GNUMAKEFLAGS: ''
-          RUBY_TESTOPTS: '--tty=no'
+          RUBY_TESTOPTS: >-
+            ${{ env.TESTS }}
+            --tty=no
           RUN_OPTS: ${{ matrix.run_opts }}
 
       - name: make test-all
